+++ ups2mqtt/rootfs/usr/src/app/ups2mqtt/web.py
@@ -9,7 +9,6 @@ import importlib.util
 import json
 import logging
 import math
-import os
 import threading
 from uuid import uuid4
 from datetime import datetime, timezone